item changess

+++ src/Items/BagUbgrade.java
@@ -39,7 +39,7 @@ public class BagUbgrade extends Item {
     }
 
     @Override
-    public int addEnergyMulti() {
+    public double addEnergyMulti() {
         return 0;
     }
