+++ src/Items/Helmet1.java
@@ -40,7 +40,7 @@ public class Helmet1 extends Item{
     }
 
     @Override
-    public int addEnergyMulti() {
+    public double addEnergyMulti() {
         return 0;
     }
