+++ src/Items/Helmet3.java
@@ -40,7 +40,7 @@ public class Helmet3 extends Item{
     }
 
     @Override
-    public int addEnergyMulti() {
+    public double addEnergyMulti() {
         return 0;
     }
