+++ src/Items/IntimidatorsThorns.java
@@ -41,7 +41,7 @@ public class IntimidatorsThorns extends Item {
     }
 
     @Override
-    public int addEnergyMulti() {
+    public double addEnergyMulti() {
         return 0;
     }
