+++ src/Items/WhiteLether.java
@@ -39,7 +39,7 @@ public class WhiteLether extends Item {
     }
 
     @Override
-    public int addEnergyMulti() {
+    public double addEnergyMulti() {
         return 0;
     }
